Fix Windows build issues and add CI workflow for cross-platform testing
- Add missing #include <windows.h> in xcvr poc_app.cpp and dummy_peer.cpp
  for SetConsoleCtrlHandler/BOOL/DWORD types
- Add -Wa,-mbig-obj flag for MinGW builds to handle COFF section limit
  with template-heavy C++23 code
- Fix -Wsign-conversion warnings for FIONBIO constant in ioctlsocket calls
- Add MinGW-w64 cross-compilation toolchain file (cmake/mingw-w64-x86_64.cmake)
- Add GitHub Actions CI workflow testing MSVC, MinGW, and Linux builds
  with conduit unit tests and xcvr client/server smoke tests
- Update .gitignore for build-* directories

Verified: cross-compilation with MinGW-w64 produces valid PE32+ executables,
native Linux build passes all 408 test cases (3379 assertions).

+++ conduit/src/transceiver/transport/tcp_client.cpp
@@ -423,7 +423,7 @@ VoidResult TcpClientTransport::Impl::try_connect() {
     #ifdef _WIN32
     {
         u_long mode = 0;
-        if (ioctlsocket(static_cast<SOCKET>(s), FIONBIO, &mode) != 0) {
+        if (ioctlsocket(static_cast<SOCKET>(s), static_cast<long>(FIONBIO), &mode) != 0) {
             close_socket(s);
             return std::unexpected(
                 CONDUIT_ERROR(ErrorCode::SocketError,
